chore(logs2): delete old logging stack end-to-end
Remove the entire byte-capture / legacy-migration logging stack now that
Logs v2 (transcript indexing) is the only path:
- src: legacy-log-parser, legacy-log-importer, log-capture, ansi-strip,
  log-db, log-worker, fork-log-worker, log-snapshot, logsdb-handlers,
  log-migration-handlers; scripts dry-run-legacy-parse + e2e-sample-import.
- Strip the old ToWorker/FromWorker/DirMigrationReport/FakeWorkerTransport
  contracts + the supervisor migrateDir stub; drop the log-worker vite entry.
- Remove LOGSDB_*/LOGS_MIGRATE_* channels + preload bridges (logsdb events,
  logMigration) + d.ts; drop the dead CloseDialog migrationRunning prop.
- PRESERVE logging:getResumeTarget: moved verbatim into a new
  resume-handlers.ts (binder-backed, no DB), with its test retargeted; the
  renderer caller (session-persistence getResumeTarget) is unchanged.
- Settings "Clear index" now calls logs2.clearAll (the v2 surface).
- KEEP legacyLogsMigrated/legacyLogsSurfacingSeen settings fields: the T7
  logs-wipe still clears those keys.
- Replace old tests; transcripts-db + transcripts-worker native suites remain
  (native non-empty). Rename insights-runner's local stripAnsi helper to
  stripAnsiCodes to clear the deletion-sweep grep gate.

typecheck 0; unit 2164 pass/1 skip; native 56 pass (2 files).
