--- src/main/ipc/log-migration-handlers.ts
@@ -1,161 +1,0 @@
-/**
- * log-migration-handlers.ts — IPC for the legacy-log migration (Phase 2b).
- * Worker-backed: NEVER imports log-db/log-worker; reaches the single live worker
- * ONLY through getLogSupervisor().migrate(...). Detection + snapshot + reclaim are
- * plain fs over <dataDir>/logs (read-only until reclaim). No default export.
- */
-import { ipcMain, BrowserWindow } from 'electron'
-import * as fs from 'fs'
-import * as path from 'path'
-import { IPC } from '../../shared/ipc-channels'
-import { getDataDirectory } from '../data-paths'
-import { getLogSupervisor } from '../logging/logging-service'
-import { snapshotLegacyLogs, isLegacyLogsFrozen, markLegacyImportComplete, readLegacyImportCompletion, reclaimLegacyLogs } from '../logging/log-snapshot'
-import { logInfo, logWarn } from '../debug-logger'
-
-// A4: module-level reentrancy guard — a double-click must not spawn two runImport loops.
-let migrationRunning = false
-// Same guard for reclaim: a double-click (or any second caller) must not start a
-// second permanent-delete pass while the first is still running.
-let reclaimRunning = false
-
-function legacyLogsDir(): string {
-  return path.join(getDataDirectory(), 'logs')
-}
-
-/** Cheap count of <label>/<id> session folders under the logs dir. Read-only,
- *  tolerant of a missing/partial tree (returns 0 rather than throwing). */
-function countSessionFolders(logsDir: string): number {
-  let labels: fs.Dirent[]
-  try {
-    labels = fs.readdirSync(logsDir, { withFileTypes: true })
-  } catch {
-    return 0
-  }
-  let count = 0
-  for (const label of labels) {
-    if (!label.isDirectory()) continue
-    const labelPath = path.join(logsDir, label.name)
-    let sessions: fs.Dirent[]
-    try {
-      sessions = fs.readdirSync(labelPath, { withFileTypes: true })
-    } catch {
-      continue
-    }
-    for (const sess of sessions) {
-      if (sess.isDirectory()) count += 1
-    }
-  }
-  return count
-}
-
-/** Current size of the SQLite log DB file (0 when absent). Used to report the
- *  DB growth a run produced. */
-function dbSizeBytes(): number {
-  try {
-    return fs.statSync(path.join(getDataDirectory(), 'logs.db')).size
-  } catch {
-    return 0
-  }
-}
-
-export function registerLogMigrationHandlers(getWindow: () => BrowserWindow | null): void {
-  ipcMain.handle(IPC.LOGS_MIGRATE_DETECT, async () => {
-    const dir = legacyLogsDir()
-    const present = fs.existsSync(dir)
-    const sessionFolders = present ? countSessionFolders(dir) : 0
-    // Persisted completion marker: lets Settings re-offer RECLAIM after an app
-    // restart (the post-run report is in-memory; the user hit a stranded,
-    // fully-armed reclaim with no UI door). Only meaningful for THIS logsDir —
-    // the reclaim gate re-checks server-side regardless.
-    const completion = readLegacyImportCompletion()
-    return {
-      present: present && sessionFolders > 0,
-      sessionFolders,
-      frozen: isLegacyLogsFrozen(),
-      completion: completion && completion.logsDir === dir ? completion : null,
-    }
-  })
-
-  ipcMain.handle(IPC.LOGS_MIGRATE_RUN, async () => {
-    // A4 guard: refuse a concurrent run rather than spawning a second import loop.
-    if (migrationRunning) throw new Error('migration already in progress')
-    migrationRunning = true
-    try {
-      const sup = getLogSupervisor()
-      if (!sup) throw new Error('logging worker not available (logging disabled?)')
-      const dir = legacyLogsDir()
-      const dbBefore = dbSizeBytes()
-      // Snapshot the detected folder count once (the tree is read-only until reclaim,
-      // so it is stable for the run) for the report's reconciliation line.
-      const detectedFolders = countSessionFolders(dir)
-      snapshotLegacyLogs()                                    // 1) one-time read-only snapshot/marker
-      // 2) Worker-internal STREAMING migration (the 16 GB fix). The worker walks +
-      // parses + imports the tree itself in bounded batches — nothing heavy ever
-      // runs on this (main) thread and nothing huge crosses the process boundary.
-      // The old in-handler parseLegacyLogs(dir) call read the WHOLE tree
-      // synchronously here and froze the app on real-size data (AppHang, Windows
-      // closed the process) — never reintroduce main-thread parsing.
-      const report = await sup.migrateDir(dir, (done, total) => {
-        try { getWindow()?.webContents.send(IPC.LOGS_MIGRATE_PROGRESS, { done, total }) } catch { /* window gone */ }
-      })
-      const dbAfter = dbSizeBytes()
-      const { unparseable, foldedPartnerDirs, noEventDirs } = report
-      // A1 [SAFETY-CRITICAL]: record import completion (the marker reclaim is later
-      // gated on) ONLY when every session either imported or was a benign
-      // already-present skip. A FAILED session (importSession threw) means its data
-      // never reached the DB; marking complete would let reclaim permanently delete
-      // that legacy folder. So on ANY failure we deliberately LEAVE THE MARKER
-      // UNWRITTEN -> reclaim stays blocked until a clean re-run. (A re-run where
-      // everything merely skips is still a clean completion: the data is in the DB.)
-      if (report.failedSessions === 0) {
-        markLegacyImportComplete({ logsDir: dir, stats: {
-          totalSessions: report.totalSessions,
-          importedSessions: report.importedSessions,
-          skippedSessions: report.skippedSessions,
-          importedEvents: report.importedEvents,
-          unparseableCount: unparseable.length,
-        } })
-      } else {
-        logWarn(`[migrate] ${report.failedSessions} session(s) failed to import; NOT marking migration complete -> reclaim stays blocked until a clean re-run`)
-      }
-      logInfo(`[migrate] imported ${report.importedSessions} sessions (${report.importedEvents} events), skipped ${report.skippedSessions}, failed ${report.failedSessions}, ${unparseable.length} unparseable`)
-      return {
-        ...report,
-        unparseable: unparseable.map((u) => ({ path: u.path, reason: u.reason, skippedLines: u.skippedLines })),
-        foldedPartnerDirs,
-        noEventDirs,
-        detectedFolders,
-        dbBytesBefore: dbBefore,
-        dbBytesAfter: dbAfter,
-      }
-    } finally {
-      migrationRunning = false   // A4: always reset, even on throw
-    }
-  })
-
-  ipcMain.handle(IPC.LOGS_MIGRATE_RECLAIM, async () => {
-    // Reentrancy guard mirrors the RUN guard. A permanent-delete double-click would
-    // otherwise start a second pass that races the first: it runs against an
-    // already-emptied tree and overwrites the real reclaim tally with a bogus
-    // near-zero result. Refuse the second call outright.
-    if (reclaimRunning) throw new Error('reclaim already in progress')
-    reclaimRunning = true
-    try {
-      // A1 [SAFETY-CRITICAL] belt-and-suspenders: never delete unless the DB actually
-      // holds imported sessions. (The frozen + completion-marker(+logsDir) gate lives
-      // in reclaimLegacyLogs.) This defends the case where an import wrote nothing.
-      const sup = getLogSupervisor()
-      if (!sup) throw new Error('refusing to reclaim: logging worker not available')
-      const rows = await sup.query('listSessions', { offset: 0, limit: 1 })
-      if (!rows.length) throw new Error('refusing to reclaim: no imported sessions in the database')
-      const res = reclaimLegacyLogs({
-        onFailure: (p, reason) => logWarn(`[migrate] reclaim failed for ${p}: ${reason}`),
-      })
-      logInfo(`[migrate] reclaimed ${res.deletedFolders} folders (${res.reclaimedBytes} bytes), ${res.failedFolders.length} failed`)
-      return res
-    } finally {
-      reclaimRunning = false
-    }
-  })
-}
