--- src/main/logging/log-worker.ts
@@ -1,548 +1,0 @@
-/**
- * log-worker.ts — logging utilityProcess worker.
- *
- * TWO PARTS:
- *  1. Pure, exported `handleWorkerMessage(db, msg, post)` — testable with a
- *     real :memory: DB and a fake `post`. All DB writes and query responses go
- *     through here; it never touches parentPort.
- *  2. Guarded bootstrap — wires `process.parentPort` to the handler, opens the
- *     DB, posts periodic health, and checkpoints WAL. The guard ensures that
- *     importing this module in a test (where parentPort is undefined) is a
- *     no-op.
- *
- * IMPORT RULES:
- *  - No `electron` import — a utilityProcess child uses `process.parentPort`,
- *    not the electron module.
- *  - Must NEVER be statically imported by main-process code. The supervisor
- *    (Task 6) forks it by FILE PATH.
- */
-import * as fs from 'fs'
-import { openLogDb } from './log-db'
-import type { LogDb } from './log-db'
-import { stripAnsi } from './ansi-strip'
-import { planLegacyGroups, streamGroup } from './legacy-log-parser'
-import type { LegacyGroup } from './legacy-log-parser'
-import type { ToWorker, FromWorker, DirMigrationReport } from './log-worker-transport'
-
-// ---------------------------------------------------------------------------
-// Internal counters — shared across all handleWorkerMessage calls made against
-// the same module instance. Tests that call handleWorkerMessage directly will
-// share these, which is fine: the tests only assert that drops are reflected,
-// not the exact cumulative total across independent test cases.
-//
-// Intentional process-lifetime singletons for health reporting. Tests must not
-// assert absolute totals across independent test cases.
-// ---------------------------------------------------------------------------
-
-let _eventsTotal = 0
-let _dropsTotal = 0
-
-/** Read the current counters (used by the bootstrap health timer). */
-export function getStats(): { eventsTotal: number; dropsTotal: number } {
-  return { eventsTotal: _eventsTotal, dropsTotal: _dropsTotal }
-}
-
-// ---------------------------------------------------------------------------
-// Pure handler — no parentPort, no side-effects beyond the DB and the post cb.
-// ---------------------------------------------------------------------------
-
-/**
- * Apply a single ToWorker message to the database and post any response via
- * `post`. `open` and `shutdown` are lifecycle messages handled by the bootstrap;
- * the pure handler ignores them (no-op) rather than throwing.
- *
- * Wrapped in try/catch — errors are reported via `post({type:'error'})` and
- * never propagate to the caller.
- */
-export function handleWorkerMessage(
-  db: LogDb,
-  msg: ToWorker,
-  post: (m: FromWorker) => void,
-): void {
-  // Extract the query id before entering the try so the catch can correlate
-  // the error back to the caller's pending promise (avoids a hung promise).
-  const queryId = msg.type === 'query' ? msg.id : undefined
-  try {
-    _handleWorkerMessage(db, msg, post)
-  } catch (err: unknown) {
-    const message = err instanceof Error ? err.message : String(err)
-    post({ type: 'error', id: queryId, message })
-  }
-}
-
-function _handleWorkerMessage(
-  db: LogDb,
-  msg: ToWorker,
-  post: (m: FromWorker) => void,
-): void {
-  switch (msg.type) {
-    // ---- open / shutdown: lifecycle — bootstrap handles these, not this fn ----
-    case 'open':
-    case 'shutdown':
-      // intentional no-op in the pure handler; lifecycle handled by the bootstrap
-      return
-
-    // ---- session-start ----
-    case 'session-start':
-      db.upsertSession(msg.meta)
-      return
-
-    // ---- session-end ----
-    case 'session-end':
-      db.finishSession(msg.sessionId, msg.ts, msg.status)
-      return
-
-    // ---- batch ----
-    case 'batch': {
-      const flat: Array<{
-        sessionId: string
-        ts: number
-        type: 'start' | 'data' | 'restart' | 'switch' | 'end'
-        raw: Buffer | Uint8Array
-        text: string
-      }> = []
-
-      for (const sess of msg.sessions) {
-        // Flatten all chunks for this session into the flat array
-        for (const chunk of sess.chunks) {
-          // Note: a multi-byte UTF-8 char split across batch boundaries may yield
-          // U+FFFD in this search-text projection only — raw bytes are stored intact
-          // so replay is unaffected; acceptable for a best-effort search index.
-          const text = stripAnsi(Buffer.from(chunk.raw).toString('utf8'))
-          flat.push({
-            sessionId: sess.sessionId,
-            ts: chunk.ts,
-            type: chunk.type,
-            raw: chunk.raw,
-            text,
-          })
-          _eventsTotal += 1
-        }
-
-        // If any bytes were dropped by the PTY ring-buffer, insert a visible
-        // gap marker so replay shows the discontinuity.
-        const dropped = sess.dropped ?? 0
-        if (dropped > 0) {
-          _dropsTotal += dropped
-          const markerText = `\n[${dropped} bytes dropped]\n`
-          flat.push({
-            sessionId: sess.sessionId,
-            ts: sess.chunks.length > 0 ? sess.chunks[sess.chunks.length - 1].ts : Date.now(),
-            type: 'data',
-            raw: Buffer.alloc(0),
-            text: markerText,
-          })
-        }
-      }
-
-      // ONE appendBatch call = one SQLite transaction for the whole message.
-      db.appendBatch(flat)
-      return
-    }
-
-    // ---- query ----
-    case 'query': {
-      const { id, kind, args } = msg
-      let rows: unknown[]
-
-      switch (kind) {
-        case 'listSessions': {
-          const opts: { offset?: number; limit?: number } = {}
-          if (typeof args.offset === 'number') opts.offset = args.offset
-          if (typeof args.limit === 'number') opts.limit = args.limit
-          rows = db.listSessions(opts)
-          break
-        }
-        case 'readEvents': {
-          const sessionId = args.sessionId as string
-          const offset = typeof args.offset === 'number' ? args.offset : 0
-          const limit = typeof args.limit === 'number' ? args.limit : 100
-          rows = db.readEvents(sessionId, { offset, limit })
-          break
-        }
-        case 'search': {
-          const query = args.query as string
-          const opts: { limit?: number } = {}
-          if (typeof args.limit === 'number') opts.limit = args.limit
-          rows = db.search(query, opts)
-          break
-        }
-        case 'prune': {
-          const ids = Array.isArray(args.ids) ? (args.ids as string[]) : []
-          const res = db.pruneSessions(ids)
-          // checkpoint after a delete so a follow-up dbBytes is honest.
-          db.checkpoint()
-          rows = [res]
-          break
-        }
-        case 'clearAll': {
-          const res = db.clearAll()
-          db.checkpoint()
-          rows = [res]
-          break
-        }
-        default: {
-          post({ type: 'error', id, message: `unknown query kind: ${kind}` })
-          return
-        }
-      }
-
-      post({ type: 'query-result', id, rows })
-      return
-    }
-
-    // ---- migrate (legacy import; runs INSIDE the single live worker) ----
-    case 'migrate': {
-      // AMENDMENT A2: wrap the ENTIRE case in try/catch. The outer
-      // handleWorkerMessage wrapper sets `id` ONLY for `query` (see :62), so a
-      // throw that escapes the per-session loop (e.g. msg.sessions not iterable,
-      // or `post` throwing) would otherwise post `{type:'error', id:undefined}`,
-      // which LogSupervisor.migrate() never matches -> the ack hangs. Here we
-      // always terminate the chunk with a migrate-error carrying msg.id.
-      try {
-        let importedSessions = 0
-        let skippedSessions = 0
-        let failedSessions = 0
-        let importedEvents = 0
-        for (const sess of msg.sessions) {
-          // Per-session try/catch so one bad session never aborts the chunk.
-          try {
-            const res = db.importSession(
-              {
-                sessionId: sess.sessionId,
-                configLabel: sess.configLabel,
-                projectCwd: sess.projectCwd,
-                accountEmail: sess.accountEmail,
-                profileId: sess.profileId,
-                provider: sess.provider,
-                startedAt: sess.startedAt,
-              },
-              sess.events.map((e) => ({ ts: e.ts, type: e.type, raw: e.raw, text: e.text })),
-            )
-            if (res.imported) {
-              importedSessions += 1
-              importedEvents += res.events
-            } else {
-              skippedSessions += 1
-            }
-          } catch (err) {
-            // Surface, never swallow. CRITICAL: a THROW means this session's data
-            // did NOT reach the DB, so it is FAILED — distinct from a benign
-            // already-present skip. The run handler must treat any failure as an
-            // incomplete migration (do not mark complete -> reclaim stays blocked),
-            // otherwise reclaim could permanently delete a folder we never imported.
-            failedSessions += 1
-            post({
-              type: 'log',
-              entry: {
-                level: 'warn',
-                message: `[migrate] session ${sess.sessionId} failed: ${err instanceof Error ? err.message : String(err)}`,
-              },
-            })
-          }
-        }
-        post({ type: 'migrate-progress', id: msg.id, importedSessions, skippedSessions, failedSessions, importedEvents })
-      } catch (err) {
-        post({ type: 'migrate-error', id: msg.id, message: err instanceof Error ? err.message : String(err) })
-      }
-      return
-    }
-
-    // ---- migrate-dir (worker-internal streaming legacy import) ----
-    case 'migrate-dir': {
-      // One directory migration at a time — a second request is refused with a
-      // correlated migrate-error rather than racing the first.
-      if (_dirMigrationRunning) {
-        post({ type: 'migrate-error', id: msg.id, message: 'a directory migration is already running' })
-        return
-      }
-      _dirMigrationRunning = true
-      // Fire-and-forget WITHIN the worker: results flow back via post() messages.
-      // runDirMigration never rejects (it posts migrate-error on fatal problems),
-      // but guard the flag reset in finally regardless.
-      void runDirMigration(db, msg, post).finally(() => {
-        _dirMigrationRunning = false
-      })
-      return
-    }
-
-    // ---- reconcile ----
-    case 'reconcile':
-      db.markRunningCrashed()
-      return
-
-    default: {
-      // Exhaustiveness guard — should never happen with correct ToWorker union.
-      const _exhaustive: never = msg
-      void _exhaustive
-      return
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// migrate-dir — worker-internal streaming legacy import (the 16 GB fix).
-//
-// The worker walks + parses the legacy tree ITSELF via the streaming parser
-// (readline — single legacy files reach ~500 MB) and imports group by group
-// through the partial-import primitives. Memory stays ~one event batch; the
-// async iteration yields to the worker's event loop between lines, so live
-// capture batches interleave naturally. A crash mid-session leaves that session
-// status='importing' (wiped + redone on re-run) and, because the run then never
-// reports failedSessions===0, the completion marker is never written and
-// reclaim stays blocked — the A1 safety chain is preserved end to end.
-// ---------------------------------------------------------------------------
-
-let _dirMigrationRunning = false
-
-/** Test seam: clear the one-at-a-time guard between native test cases. */
-export function __resetDirMigrationForTest(): void {
-  _dirMigrationRunning = false
-}
-
-const PROGRESS_INTERVAL_MS = 200
-
-/** Import ONE group via the streaming parser. Returns its contribution to the
- *  report. Throws on a DB failure — the caller tallies the session as FAILED. */
-async function importGroup(
-  db: LogDb,
-  group: LegacyGroup,
-  batchBytes: number | undefined,
-): Promise<{
-  imported: boolean
-  events: number
-  unparseable: DirMigrationReport['unparseable']
-  foldedPartnerDirs: number
-  noEventDirs: number
-}> {
-  let established = false
-  let eventsImported = 0
-  let tallies: { unparseable: DirMigrationReport['unparseable']; foldedPartnerDirs: number; noEventDirs: number } = {
-    unparseable: [],
-    foldedPartnerDirs: 0,
-    noEventDirs: 0,
-  }
-  let imported = false
-
-  for await (const msg of streamGroup(group, batchBytes)) {
-    if (msg.kind === 'meta') {
-      const meta: Parameters<LogDb['beginPartialImport']>[0] = {
-        sessionId: msg.meta.sessionId,
-        configLabel: msg.meta.configLabel,
-        provider: msg.meta.provider,
-        startedAt: msg.firstTs,
-      }
-      if (msg.meta.accountEmail !== undefined) meta.accountEmail = msg.meta.accountEmail
-      if (msg.meta.profileId !== undefined) meta.profileId = msg.meta.profileId
-      db.beginPartialImport(meta)
-      established = true
-    } else if (msg.kind === 'events') {
-      db.appendBatch(
-        msg.events.map((e) => ({
-          sessionId: group.baseId,
-          ts: e.ts,
-          type: e.type,
-          raw: Buffer.from(e.data ?? '', 'utf8'),
-          text: stripAnsi(e.data ?? ''),
-        })),
-      )
-      eventsImported += msg.events.length
-    } else {
-      tallies = { unparseable: msg.unparseable, foldedPartnerDirs: msg.foldedPartnerDirs, noEventDirs: msg.noEventDirs }
-      if (msg.hadEvents && established) {
-        db.completePartialImport(group.baseId, { startedAt: msg.minTs, endedAt: msg.maxTs })
-        imported = true
-      }
-    }
-  }
-
-  return { imported, events: eventsImported, ...tallies }
-}
-
-/**
- * Drive a full directory migration. NEVER rejects: fatal problems post
- * migrate-error; per-session failures are tallied as failedSessions (their
- * 'importing' marker survives for the re-run) and never abort the run.
- * Exported for the native test, which awaits it directly for determinism.
- */
-export async function runDirMigration(
-  db: LogDb,
-  msg: { id: number; logsDir: string; batchBytes?: number },
-  post: (m: FromWorker) => void,
-): Promise<void> {
-  try {
-    const groups = planLegacyGroups(msg.logsDir)
-    const total = groups.length
-    const report: DirMigrationReport = {
-      totalSessions: 0,
-      importedSessions: 0,
-      skippedSessions: 0,
-      failedSessions: 0,
-      importedEvents: 0,
-      unparseable: [],
-      foldedPartnerDirs: 0,
-      noEventDirs: 0,
-    }
-
-    let done = 0
-    let lastProgressAt = 0
-    for (const group of groups) {
-      try {
-        // Pre-skip WITHOUT parsing: a session already complete in the DB (re-run,
-        // or live-captured). A dead partial (status='importing') falls through to
-        // a full wipe + re-import.
-        const state = db.getSessionImportState(group.baseId)
-        if (state && state.eventCount > 0 && state.status !== 'importing') {
-          report.totalSessions += 1
-          report.skippedSessions += 1
-          // Reconciliation approximation for unparsed groups: every extra member
-          // dir is attributed as folded (cannot split folded/noEvent without
-          // parsing). Keeps detectedFolders === total + folded + noEvent exact.
-          report.foldedPartnerDirs += group.members.length - 1
-        } else {
-          const res = await importGroup(db, group, msg.batchBytes)
-          for (const u of res.unparseable) report.unparseable.push(u)
-          report.foldedPartnerDirs += res.foldedPartnerDirs
-          report.noEventDirs += res.noEventDirs
-          if (res.imported) {
-            report.totalSessions += 1
-            report.importedSessions += 1
-            report.importedEvents += res.events
-          }
-          // !res.imported && noEventDirs counted above — an all-malformed group
-          // contributes no session row and no totalSessions entry (old semantics).
-        }
-      } catch (err) {
-        // The session's data did NOT fully reach the DB (its row stays
-        // status='importing'). Tally as FAILED — the handler will refuse to write
-        // the completion marker, keeping reclaim blocked. Attribute the group's
-        // extra member dirs as folded so the report still reconciles.
-        report.totalSessions += 1
-        report.failedSessions += 1
-        report.foldedPartnerDirs += group.members.length - 1
-        post({
-          type: 'log',
-          entry: {
-            level: 'warn',
-            message: `[migrate-dir] session ${group.baseId} failed: ${err instanceof Error ? err.message : String(err)}`,
-          },
-        })
-      }
-
-      done += 1
-      const now = Date.now()
-      if (now - lastProgressAt >= PROGRESS_INTERVAL_MS || done === total) {
-        post({ type: 'migrate-dir-progress', id: msg.id, done, total })
-        lastProgressAt = now
-      }
-      // Explicit yield between groups so queued live-capture messages are handled
-      // even when a group parsed entirely from cache without awaiting I/O.
-      await new Promise((resolve) => setImmediate(resolve))
-    }
-
-    post({ type: 'migrate-dir-done', id: msg.id, report })
-  } catch (err) {
-    post({ type: 'migrate-error', id: msg.id, message: err instanceof Error ? err.message : String(err) })
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Guarded utilityProcess bootstrap.
-//
-// `process.parentPort` is only defined when running inside a utilityProcess
-// child (Electron). When this module is imported in tests, parentPort is
-// undefined and the block below is skipped entirely — safe to import freely.
-// ---------------------------------------------------------------------------
-
-// Top-level fs import is fine: Node built-in, loaded in any Node/worker context.
-
-const parentPort = (process as any).parentPort as
-  | {
-      on(event: 'message', handler: (e: { data: unknown }) => void): void
-      postMessage(msg: FromWorker): void
-    }
-  | undefined
-
-if (parentPort) {
-  // Single mutable state for the bootstrap — both the message handler and the
-  // health timer close over these variables.
-  let db: LogDb | undefined
-  let openedDbPath: string | undefined
-
-  const post = (m: FromWorker) => parentPort!.postMessage(m)
-
-  // ---- periodic health report (every 10 s) ----
-  // dbBytes: read the DB file size when available; :memory: reports 0.
-  // Declared before the message handler so shutdown can call clearInterval().
-  const HEALTH_INTERVAL_MS = 10_000
-
-  const healthTimer = setInterval(() => {
-    let dbBytes = 0
-    if (openedDbPath && openedDbPath !== ':memory:') {
-      try {
-        dbBytes = fs.statSync(openedDbPath).size
-      } catch {
-        // DB file might not be flushed yet — ignore
-      }
-    }
-    const { eventsTotal, dropsTotal } = getStats()
-    post({
-      type: 'health',
-      inFlight: 0, // synchronous worker — no async in-flight ops
-      eventsTotal,
-      dropsTotal,
-      dbBytes,
-    })
-  }, HEALTH_INTERVAL_MS).unref()
-
-  parentPort.on('message', (e) => {
-    // Defense-in-depth: the lifecycle branches (open/shutdown) and the pre-dispatch
-    // guards below run OUTSIDE handleWorkerMessage's own try/catch. An exception
-    // escaping here would become an uncaught exception that exits the worker
-    // process; a deterministic poison message could then drive a restart loop
-    // straight to permanent degrade. Catch + report instead of crashing.
-    try {
-      const msg = e.data as ToWorker
-
-      // ---- open: lifecycle ----
-      if (msg.type === 'open') {
-        try {
-          db = openLogDb(msg.dbPath)
-          openedDbPath = msg.dbPath
-          post({ type: 'ready' })
-        } catch (err: unknown) {
-          post({
-            type: 'error',
-            message: `failed to open DB: ${err instanceof Error ? err.message : String(err)}`,
-          })
-        }
-        return
-      }
-
-      // ---- shutdown: lifecycle ----
-      if (msg.type === 'shutdown') {
-        clearInterval(healthTimer)
-        db?.close()
-        db = undefined
-        openedDbPath = undefined
-        // Let the process exit naturally; the supervisor will await it.
-        return
-      }
-
-      // ---- all other messages require an open DB ----
-      if (!db) {
-        post({ type: 'error', message: `worker received ${msg.type} before open` })
-        return
-      }
-
-      handleWorkerMessage(db, msg, post)
-    } catch (err: unknown) {
-      post({ type: 'error', message: `worker message handling failed: ${err instanceof Error ? err.message : String(err)}` })
-    }
-  })
-
-  // TODO (Task 6/8): WAL checkpoint cadence. Preferred approach: add a
-  // `checkpoint()` method to LogDb (it owns the connection) and call it from
-  // the periodic health timer or a separate interval. Avoids reaching into raw
-  // SQLite from here. Deferred to avoid scope-creep in Task 5.
-}
